+++ src/routes/commercialRegistrations/find.ts
@@ -1,50 +1,8 @@
 import type { Request, Response } from "express";
 import { businesses } from "../../fixtures/businesses";
-import {
-  SearchQuery,
-  SearchByNameResponseBody,
-  findQuery,
-  ModelNotFoundError,
-  Business,
-} from "./types";
-
-/**
- * @see {@link https://docs.solarisgroup.com/api-reference/onboarding/businesses/#tag/Business-Registrations/paths/~1v1~1commercial_registrations~1search_by_name/get}
- * Note: This endpoint may be used free of charge.
- *
- * Returns some business as if it exists
- * Returns a mock business if more than 2 characters are provided for country and name
- */
-export const searchByName = async (
-  req: Request<{}, {}, {}, SearchQuery>,
-  res: Response<SearchByNameResponseBody | ModelNotFoundError>
-) => {
-  const { country = "DE", name } = req.query;
-  const foundBusiness = businesses.find(
-    (business) => business.name === name && business.address.country === country
-  );
-  const highEffort = String(country).length + String(name).length > 2;
-
-  if (foundBusiness) {
-    return res.status(200).send(foundBusiness);
-  } else if (highEffort) {
-    const mockBusiness: Business = {
-      ...businesses[0],
-      name,
-      address: { ...businesses[0].address, country },
-    };
-    return res.status(200).send(mockBusiness);
-  } else {
-    const errorResponse: ModelNotFoundError = {
-      title: "Model Not Found",
-      status: "404",
-      code: "model_not_found",
-      detail: "someString",
-      id: "someString",
-    };
-    return res.status(404).send(errorResponse);
-  }
-};
+import { ModelNotFoundError } from "./types/modelNotFoundError";
+import { Business } from "./types/business";
+import { findQuery } from "./types/findQuery";
 
 /**
  * @see {@link https://docs.solarisgroup.com/api-reference/onboarding/businesses/#tag/Business-Registrations/paths/~1v1~1commercial_registrations~1find/get}
