+++ src/routes/commercialRegistrations/types/country.ts
@@ -1,0 +1,1 @@
+export type Country = string | "DE";
