+++ src/routes/commercialRegistrations/types/findQuery.ts
@@ -1,0 +1,7 @@
+import { Country } from "./country";
+
+export type findQuery = {
+  registration_number: string;
+  registration_issuer: string;
+  country: Country;
+};
