+++ src/routes/commercialRegistrations/types/legalForm.ts
@@ -1,0 +1,1 @@
+export type LegalForm = string | "GMBH";
