+++ src/routes/commercialRegistrations/types/legalRepresentative.ts
@@ -1,0 +1,4 @@
+export type LegalRepresentative = {
+  first_name: string;
+  last_name: string;
+};
