+++ src/routes/commercialRegistrations/types/modelNotFoundError.ts
@@ -1,0 +1,7 @@
+export type ModelNotFoundError = {
+  title: string | "Model Not Found";
+  status: string | "404";
+  id: string;
+  detail: string;
+  code: string | "model_not_found";
+};
