+++ src/routes/commercialRegistrations/types/searchQuery.ts
@@ -1,0 +1,4 @@
+export type SearchQuery = {
+  country: string;
+  name: string;
+};
