+++ src/routes/commercialRegistrations/types/searchResponseBody.ts
@@ -1,0 +1,8 @@
+export type SearchResponseBody = {
+  name: string;
+  registration_number: string;
+  /**
+   * Only required for companies in Germany
+   */
+  registration_issuer?: string;
+};
